% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrid.R
\name{aggrid}
\alias{aggrid}
\title{Display a data frame in aggrid table}
\usage{
aggrid(
  df,
  columnDefs = NULL,
  suppressFieldDotNotation = TRUE,
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{df}{\link{data.frame} | \link{matrix}: data frame or matrix.}

\item{columnDefs}{\link{list}: List of Column / Column Group definitions. See \url{https://www.ag-grid.com/javascript-data-grid/column-definitions/}.}

\item{suppressFieldDotNotation}{\link{logical}: If TRUE, then dots in field names (e.g. 'address.firstLine') are not treated as deep references. Allows you to use dots in your field name if you prefer. \code{TRUE} or \code{FALSE}. \url{https://www.ag-grid.com/javascript-data-grid/grid-options/#reference-columns-suppressFieldDotNotation}.}

\item{width}{\link{integer}, \link{character}: Width in pixels (optional, defaults to automatic sizing)}

\item{height}{\link{integer}, \link{character}: Height in pixels (optional, defaults to automatic sizing)}

\item{elementId}{\link{character}: An id for the widget (a random character by default).}
}
\description{
Create a HTML widget using the ag-grid-react library
}
\section{Returns}{

\preformatted{
   Type: html widget
}
}

\examples{
\dontrun{
aggrid::aggrid(iris)
}
}
